Expose injection of custom JavaScript files
