+++ turbolinks/src/main/java/com/basecamp/turbolinks/TurbolinksHelper.java
@@ -4,8 +4,6 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.MutableContextWrapper;
 import android.os.Handler;
-import android.util.Base64;
-import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
@@ -15,14 +13,10 @@ import com.google.gson.GsonBuilder;
 
 import org.apache.commons.lang3.StringUtils;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 import java.net.URL;
 
 class TurbolinksHelper {
-    private static String scriptInjectionFormat = "(function(){var parent = document.getElementsByTagName('head').item(0);var script = document.createElement('script');script.type = 'text/javascript';script.innerHTML = window.atob('%s');parent.appendChild(script);return true;})()";
-
     // ---------------------------------------------------
     // Package public
     // ---------------------------------------------------
@@ -55,47 +49,6 @@ class TurbolinksHelper {
             return uri.toURL().toString();
         } catch (Exception e) {
             return null;
-        }
-    }
-
-    /**
-     * <p>Gets the base64-encoded string of a local asset file (typically a Javascript or HTML file)</p>
-     *
-     * @param context  An activity context.
-     * @param filePath Local file path relative to the main/src directory.
-     * @return A base-64 encoded string of the file contents.
-     * @throws IOException Typically if a file cannot be found or read in.
-     */
-    static String getContentFromAssetFile(Context context, String filePath) throws IOException {
-        InputStream inputStream = context.getAssets().open(filePath);
-        byte[] buffer = new byte[inputStream.available()];
-        inputStream.read(buffer);
-        inputStream.close();
-        return Base64.encodeToString(buffer, Base64.NO_WRAP);
-    }
-
-    /**
-     * <p>Injects Javascript into the webView.</p>
-     *
-     * @param turbolinksSession The TurbolinksSession.
-     * @param context           Any Android context.
-     * @param webView           The shared webView.
-     */
-    static void injectTurbolinksBridge(final TurbolinksSession turbolinksSession, Context context, WebView webView) {
-        try {
-            String script = TurbolinksHelper.getContentFromAssetFile(context, "js/turbolinks_bridge.js");
-            String jsCall = String.format(scriptInjectionFormat, script);
-
-            webView.evaluateJavascript(jsCall, new ValueCallback<String>() {
-                @Override
-                public void onReceiveValue(String s) {
-                    if (turbolinksSession != null) {
-                        turbolinksSession.turbolinksBridgeInjected = Boolean.parseBoolean(s);
-                    }
-                }
-            });
-        } catch (IOException e) {
-            TurbolinksLog.e("Error injecting script file into webview: " + e.toString());
         }
     }
